requirements.txt
pyyaml
numpy
onnx
onnx_graphsurgeon
